// development/OnBoarding/OnBoarding/OnBoarding.Web.Fnz/src/typings/ApiClient.ts
export class HomeDashboard implements IHomeDashboard {
    dashboard?: Dashboard[] | undefined;
    fse?: DashboardFse | undefined;

    constructor(data?: IHomeDashboard) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            if (data["Dashboard"] && data["Dashboard"].constructor === Array) {
                this.dashboard = [] as any;
                for (let item of data["Dashboard"])
                    this.dashboard!.push(Dashboard.fromJS(item));
            }
            this.fse = data["Fse"] ? DashboardFse.fromJS(data["Fse"]) : <any>undefined;
        }
    }

    static fromJS(data: any): HomeDashboard {
        data = typeof data === 'object' ? data : {};
        let result = new HomeDashboard();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (this.dashboard && this.dashboard.constructor === Array) {
            data["Dashboard"] = [];
            for (let item of this.dashboard)
                data["Dashboard"].push(item.toJSON());
        }
        data["Fse"] = this.fse ? this.fse.toJSON() : <any>undefined;
        return data; 
    }
}

export interface IHomeDashboard {
    dashboard?: Dashboard[] | undefined;
    fse?: DashboardFse | undefined;
}

export class Dashboard implements IDashboard {
    projectId!: string;
    teamId?: string | undefined;
    roleId?: string | undefined;
    projectName?: string | undefined;
    teamName?: string | undefined;
    roleName?: string | undefined;
    count!: number;
    completedCount!: number;
    associateId?: string | undefined;
    associateName?: string | undefined;
    podCompletionPercentage!: number;
    fse?: boolean | undefined;
    fseEligibility?: boolean | undefined;
    desiredLevel!: number;

    constructor(data?: IDashboard) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.projectId = data["ProjectId"];
            this.teamId = data["TeamId"];
            this.roleId = data["RoleId"];
            this.projectName = data["ProjectName"];
            this.teamName = data["TeamName"];
            this.roleName = data["RoleName"];
            this.count = data["Count"];
            this.completedCount = data["CompletedCount"];
            this.associateId = data["AssociateId"];
            this.associateName = data["AssociateName"];
            this.podCompletionPercentage = data["PodCompletionPercentage"];
            this.fse = data["Fse"];
            this.fseEligibility = data["FseEligibility"];
            this.desiredLevel = data["DesiredLevel"];
        }
    }

    static fromJS(data: any): Dashboard {
        data = typeof data === 'object' ? data : {};
        let result = new Dashboard();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ProjectId"] = this.projectId;
        data["TeamId"] = this.teamId;
        data["RoleId"] = this.roleId;
        data["ProjectName"] = this.projectName;
        data["TeamName"] = this.teamName;
        data["RoleName"] = this.roleName;
        data["Count"] = this.count;
        data["CompletedCount"] = this.completedCount;
        data["AssociateId"] = this.associateId;
        data["AssociateName"] = this.associateName;
        data["PodCompletionPercentage"] = this.podCompletionPercentage;
        data["Fse"] = this.fse;
        data["FseEligibility"] = this.fseEligibility;
        data["DesiredLevel"] = this.desiredLevel;
        return data; 
    }
}

export interface IDashboard {
    projectId: string;
    teamId?: string | undefined;
    roleId?: string | undefined;
    projectName?: string | undefined;
    teamName?: string | undefined;
    roleName?: string | undefined;
    count: number;
    completedCount: number;
    associateId?: string | undefined;
    associateName?: string | undefined;
    podCompletionPercentage: number;
    fse?: boolean | undefined;
    fseEligibility?: boolean | undefined;
    desiredLevel: number;
}

export class DashboardFse implements IDashboardFse {
    completedFse!: number;
    fseCount!: number;

    constructor(data?: IDashboardFse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.completedFse = data["CompletedFse"];
            this.fseCount = data["FseCount"];
        }
    }

    static fromJS(data: any): DashboardFse {
        data = typeof data === 'object' ? data : {};
        let result = new DashboardFse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["CompletedFse"] = this.completedFse;
        data["FseCount"] = this.fseCount;
        return data; 
    }
}

export interface IDashboardFse {
    completedFse: number;
    fseCount: number;
}